fix: Fix incorret size_t usage
Declaring n_aug_ as size_t caused an error in subtraction.

+++ src/ukf.h
@@ -28,10 +28,10 @@ public:
   bool use_radar_;
 
   // State dimension
-  size_t n_x_;
+  int n_x_;
 
   // Augmented state dimension
-  size_t n_aug_;
+  int n_aug_;
 
   // Sigma point spreading parameter
   double lambda_;
